+++ modules/main/src/main/java/com/nisovin/shopkeepers/config/Settings.java
@@ -55,7 +55,7 @@ public class Settings extends Config {
 	/*
 	 * General Settings
 	 */
-	public static int configVersion = 8;
+	public static int configVersion = 9;
 	public static boolean debug = false;
 	// See DebugOptions for all available options.
 	public static List<String> debugOptions = new ArrayList<>(0);