+++ modules/main/src/main/java/com/nisovin/shopkeepers/config/migration/ConfigMigrations.java
@@ -22,7 +22,8 @@ public class ConfigMigrations {
 			new ConfigMigration5(),
 			new ConfigMigration6(),
 			new ConfigMigration7(),
-			new ConfigMigration8()
+			new ConfigMigration8(),
+			new ConfigMigration9()
 	);
 
 	public static int getLatestVersion() {