+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Date format: (YYYY-MM-DD)
 ## v2.23.2 (TBA)
 ### Supported MC versions: 1.21.3, 1.21.1, 1.21, 1.20.6
 
+* Fix: Config sound effects fail to play on Spigot 1.21.3. Add config migration from sound enum names to namespaced keys.
 
 ## v2.23.1 (2024-11-03)
 ### Supported MC versions: 1.21.3, 1.21.1, 1.21, 1.20.6